Extract code from SeriesDetailViewController
