+++ Example/Example/SeriesDetailCollectionViewDataSource.swift
@@ -1,0 +1,34 @@
+//
+//  SeriesDetailCollectionViewDataSource.swift
+//  Example
+//
+//  Created by Pedro Vicente Gomez on 23/12/15.
+//  Copyright © 2015 GoKarumi S.L. All rights reserved.
+//
+
+import Foundation
+import BothamUI
+
+class SeriesDetailCollectionViewDataSource<U, V: BothamViewCell where U == V.ItemType> : BothamCollectionViewDataSource<U, V> {
+
+    var seriesHeader: Series?
+
+    func collectionView(collectionView: UICollectionView,
+        viewForSupplementaryElementOfKind kind: String,
+        atIndexPath indexPath: NSIndexPath) -> UICollectionReusableView {
+            if kind == UICollectionElementKindSectionHeader {
+                let headerView =
+                collectionView.dequeueReusableSupplementaryViewOfKind(kind,
+                    withReuseIdentifier: "SeriesDetailCollectionHeaderViewReusableIdentifier",
+                    forIndexPath: indexPath)
+                    as! SeriesDetailCollectionHeaderView
+                if let header = self.seriesHeader {
+                    headerView.configureForItem(header)
+                }
+                return headerView
+
+            }
+            assert(false, "Unexpected element kind")
+    }
+
+}
