+++ Example/Example/SeriesDetailViewController.swift
@@ -16,11 +16,9 @@ private struct Config {
     static let footerHeight = 20
     static let cellMargin = 20
 }
-class SeriesDetailViewController : ExampleViewController, BothamCollectionViewController, SeriesDetailUI, UICollectionViewDelegate, UICollectionViewDataSource {
-
+class SeriesDetailViewController : ExampleViewController, BothamCollectionViewController, SeriesDetailUI {
 
     var dataSource: BothamCollectionViewDataSource<Comic, ComicCollectionViewCell>!
-    var header: Series?
 
     @IBOutlet weak var collectionView: UICollectionView!
 
@@ -38,39 +36,9 @@ class SeriesDetailViewController : ExampleViewController, BothamCollectionViewCo
         super.viewWillDisappear(animated)
     }
 
-    func collectionView(
-        collectionView: UICollectionView,
-        numberOfItemsInSection section: Int) -> Int {
-            return self.dataSource.collectionView(collectionView, numberOfItemsInSection: section)
-    }
-
-    func collectionView(
-        collectionView: UICollectionView,
-        cellForItemAtIndexPath indexPath: NSIndexPath) -> UICollectionViewCell {
-            return dataSource.collectionView(collectionView,
-                cellForItemAtIndexPath: indexPath)
-    }
-
-    func collectionView(collectionView: UICollectionView,
-        viewForSupplementaryElementOfKind kind: String,
-        atIndexPath indexPath: NSIndexPath) -> UICollectionReusableView {
-            if kind == UICollectionElementKindSectionHeader {
-                let headerView =
-                collectionView.dequeueReusableSupplementaryViewOfKind(kind,
-                    withReuseIdentifier: "SeriesDetailCollectionHeaderViewReusableIdentifier",
-                    forIndexPath: indexPath)
-                    as! SeriesDetailCollectionHeaderView
-                if let header = self.header {
-                    headerView.configureForItem(header)
-                }
-                return headerView
-
-            }
-            assert(false, "Unexpected element kind")
-    }
 
     func configureHeader(series: Series) {
-        header = series
+        (dataSource as? SeriesDetailCollectionViewDataSource)?.seriesHeader = series
     }
 
     private func configureNavigationBar() {
@@ -81,21 +49,21 @@ class SeriesDetailViewController : ExampleViewController, BothamCollectionViewCo
     }
 
     private func configureCollectionView() {
-        dataSource = BothamCollectionViewDataSource<Comic, ComicCollectionViewCell>()
+        dataSource = SeriesDetailCollectionViewDataSource()
         let navBarHeight = navigationController?.navigationBar.frame.height ?? 0
         let statusBarHeight = UIApplication.sharedApplication().statusBarFrame.size.height
         let topInset = navBarHeight + statusBarHeight
         collectionView.accessibilityLabel = "ComicsCollectionView"
         collectionView.contentInset = UIEdgeInsetsMake(-topInset, 0, 0, 0)
-        collectionView.dataSource = self
+        collectionView.dataSource = dataSource
         let layout = UICollectionViewFlowLayout()
+        layout.headerReferenceSize = CGSize(width: view.frame.width, height: CGFloat(Config.headerHeight))
         layout.itemSize = CGSize(
             width: view.frame.width / CGFloat(Config.numberOfColumns),
             height: CGFloat(Config.cellHeight))
         layout.minimumInteritemSpacing = 0
         layout.minimumLineSpacing = CGFloat(Config.cellMargin)
         collectionView.setCollectionViewLayout(layout, animated: true)
-        layout.headerReferenceSize = CGSize(width: view.frame.width, height: CGFloat(Config.headerHeight))
     }
 
 }
